requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.2
django-environ==0.8.1
django-heroku==0.3.1
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
environ==1.0
gunicorn==20.1.0
phonenumbers==8.12.42
psycopg2==2.9.3
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
whitenoise==6.0.0
xlwt==1.3.0
